pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "build", "twine"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = ["-s", "--import-mode=importlib"]
pythonpath = [".", "src", "tests"]
testpaths = ["tests/unit"]

[project]
name = "reqstool-python-decorators"
dynamic = ["version"]
authors = [{ name = "reqstool" }]
description = "Reqstool Python Decorators"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
]
dependencies = ["ruamel.yaml>=0.18.5"]

[project.urls]
Homepage = "https://reqstool.github.io"
Source = "https://github.com/reqstool/reqstool-python-decorators.git"
Documentation = "https://github.com/reqstool/reqstool-python-decorators.git"

[project.scripts]
# section needed for poetry convertion

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.envs.dev]
dependencies = [
    "pytest==8.4.1",
    "pytest-sugar==1.0.0",
    "pytest-cov==6.2.1",
    "black==25.1.0",
    "flake8==7.3.0",
    "flake8-pyproject==1.2.3",
]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.flake8]
ignore = ["W503"]
max-line-length = 120

max-complexity = 10
